Set the exact version of emsdk
Previously we installed a specific version, but we were using the latest
git version of emsdk, so things broke when upstream updated their nodejs
version. By setting the same version of both the git repository and the
installed package, we ensure that things do not break unless we
explicitly upgrade.